With the rapidly expanding increase in the number of new diagnostics and commands (and more planned), I am attempting to establish a standard now which will make some future transitions and plans easier.
Although some commands must create new GData data structures due to the new tag system (for example, so that the computed velocity vector has the appropriate tag and labels, and does not overwrite any input GData data structures), as much as possible, I am seeking to keep the GData data structure the key piece of manipulable data (especially in new commands that use stack=True and thus just push the resulting computation into the GData data structure) and NOT do numpy array manipulations in postgkyl commands.

Most importantly, I am setting a standard that any and all operations added to the postgkyl diagnostics folder should take GData data structures as inputs, NOT numpy arrays. The functions in diagnostics can call methods such as .getValues() to obtain numpy arrays to manipulate, but there is significant value in what is passed to the diagnostics functions being GData data structures. The diagnostics can then use metadata (such as .charge in current accumulation) and critically, will allow us to have a single function which parses whether the input GData data structure is p=0 (finite volume) or p>0 (DG) data, and change the operation if it detects that the data is DG data (for example, using weak operations in the computation of the magnitude squared of a vector, or the rotation operations).